// src/app/_enum/sideBar1MenuAuthority.enum.ts
export enum SideBar1MenuAuthorityEnum {
    ACCESS_DASHBOARD = "menu:dashboard",
    ACCESS_PATIENT = "menu:patient",
    ACCESS_ADMISSION = "menu:admission",
    ACCESS_CAISSE = "menu:caisse", 
    ACCESS_INSURANCE_INVOICE = "menu:insurance_invoice",
    ACCESS_CONSTANT_TAKING = "menu:constant_taking",
    ACCESS_WAITING_ROOM = "menu:waiting_room",
    ACCESS_PHARMACY = "menu:pharmacy",
    ACCESS_CUSTOMER_SALE = "menu:customer_sale",
    ACCESS_HOSPITALISATION ="menu:hospitalisation",
    ACCESS_STATISTICS = "menu:statistics",
    ACCESS_LABORATORY= "menu:laboratoire"    
}